requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==4.0.4
djangorestframework==3.13.1
hashids==1.3.1
idna==3.3
Pillow==9.1.1
psycopg2-binary==2.8.4
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
stripe==3.1.0
urllib3==1.26.9
